Cambios 27/03/2023 libreria chars.js y laravel sptie roles y permisos

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('CI')->unique();
+            $table->integer('CI')->unique();
             $table->string('name');
             $table->string('email');
             $table->string('username');
